Add "similar to" concept
We want to capture similarity relationships between many entities.  Some of these are structured from ontologies, such as the included terms from CHEBI, but others may be based on some numerical calculation of similarity (e.g. owlsim or chemical structural similarity)

In this PR, similar to is a symmetric property, with mappings to RO, SO, and CHEBI.  It is directly subclassed from related to, and it is the parent of model_of and homologous_to.