remove slot_uri from subclass of
subclass of is a "related to" not an association slot. 
So I think it should not have a defined slot_uri (just as e.g. superclass of does not).

The reason this is a problem is that for all other predicates, I can use the toolkit get_element().slot_uri to get a predicate like
"biolink:superclass_of".  But if I do that with subclass I get "rdfs:SubclassOf" which annoys things that are expecting to find biolink:whatever.