Update biolink-model.yaml
In reviewing ROBOKOP for non-biolink edge properties, we found that  4 sources (BindingDB, DrugCentral, Pharos, and GToPdb) needed  to express the relationship between a chemical and a protein numerically, using things like pIC50 values.  So we could add slots like pIC50, pEC50, etc.  But it turns out that there are a lot of these.  Currently in robokop there are about 25 (only a few of the most common are included in this PR).

So in robokop we have a strategy of taking something like a pIC50 of 8 and turning it into:
```
affinity: 8
affinity_parameter: pIC50
```